requirements.txt
asgiref>=3.2.3
certifi>=2019.11.28
chardet>=3.0.4
dj-database-url>=0.5.0
Django>=3.0.2
django-cors-headers>=3.2.1
django-postgres-copy==2.3.1
python-decouple==3.1
django-heroku>=0.3.1
django-allauth>=0.41.0
future>=0.18.2
gunicorn>=20.0.4
idna>=2.8
joblib>=0.14.1
lxml>=4.4.2
psycopg2>=2.8.4
python-dateutil>=2.8.1
python-whois>=0.7.2
pytz>=2019.3
requests>=2.22.0
six>=1.14.0
sklearn
sqlparse>=0.3.0
urllib3>=1.25.8
whitenoise>=5.0.1
xmltodict>=0.12.0
django-reset-migrations==0.4.0
scikit_learn==0.22.2.post1
